LayoutTests/http/tests/security/xss-DENIED-assign-location-href-javascript-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/localPage.html from frame with URL http://127.0.0.1:8000/security/xss-DENIED-assign-location-href-javascript.html. Domains, protocols and ports must match.

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8000/security/resources/localPage.html from frame with URL http://127.0.0.1:8000/security/xss-DENIED-assign-location-href-javascript.html. Domains, protocols and ports must match.



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
You can see the contents of this file.
